Add files via upload
This took so much longer for me to complete and so much more code than I anticipated. Compared to the check_in, I wanted to change the pictograms to include the full list and be identified by their number, because not all pictograms show the title but they do show the number. I also wanted to add input validations. I spent hours looking online to figure out how to make those changes and make it work. Half of the added code is the GUI with tkinter. I've never used this before so it took me several days and lots of breaks to get it working. Thankfully, the final code WORKS.